+++ RelationalAI.Examples/CreateDatabase.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.CommandLine;
 using System.CommandLine.NamingConventionBinder;
+using System.Threading.Tasks;
 using RelationalAI;
 
 
@@ -32,13 +33,13 @@ namespace RelationalAI.Examples
             return cmd;
         }
 
-        private static void Run(string database, string engine, string profile = "default")
+        private static async Task Run(string database, string engine, string profile = "default")
         {
             Dictionary<string, object> config = Config.Read("", profile);
             Client.Context context = new Client.Context(config);
             Client client = new Client(context);
             Console.WriteLine("Creating Database: "+ database);
-            Console.WriteLine(client.CreateDatabase(database, engine));
+            Console.WriteLine(await client.CreateDatabaseAsync(database, engine));
         }
 
     }