+++ RelationalAI.Examples/GetTransactionMetadata.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.CommandLine;
 using System.CommandLine.NamingConventionBinder;
+using System.Threading.Tasks;
 
 namespace RelationalAI.Examples
 {
@@ -25,12 +26,13 @@ namespace RelationalAI.Examples
             return cmd;
         }
 
-        private static void Run(string id, string profile = "default")
+        private static async Task Run(string id, string profile = "default")
         {
             Dictionary<string, object> config = Config.Read("", profile);
             Client.Context context = new Client.Context(config);
             Client client = new Client(context);
-            var metadata = client.GetTransactionMetadata(id);
+
+            var metadata = await client.GetTransactionMetadataAsync(id);
             Console.WriteLine(metadata);
         }
     }