+++ RelationalAI.Examples/GetTransactionProblems.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.CommandLine;
 using System.CommandLine.NamingConventionBinder;
+using System.Threading.Tasks;
 
 namespace RelationalAI.Examples
 {
@@ -25,12 +26,12 @@ namespace RelationalAI.Examples
             return cmd;
         }
 
-        private static void Run(string id, string profile = "default")
+        private static async Task Run(string id, string profile = "default")
         {
             Dictionary<string, object> config = Config.Read("", profile);
             Client.Context context = new Client.Context(config);
             Client client = new Client(context);
-            var problems = client.GetTransactionProblems(id);
+            var problems = await client.GetTransactionProblemsAsync(id);
             foreach(var problem in problems)
             {
                 Console.WriteLine(problem);