+++ RelationalAI.Test/ExecuteAsync.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Collections.Generic;
 using Relationalai.Protocol;
 using Xunit;
+using System.Threading.Tasks;
 
 namespace RelationalAI.Test
 {
@@ -12,15 +13,15 @@ namespace RelationalAI.Test
         public static string Dbname = $"csharp-sdk-{UUID}";
         public static string EngineName = $"csharp-sdk-{UUID}";
         [Fact]
-        public void ExecuteAsyncTest()
+        public async Task ExecuteAsyncTest()
         {
             Client client = CreateClient();
 
-            client.CreateEngineWait(EngineName);
-            client.CreateDatabase(Dbname, EngineName);
+            await client.CreateEngineWaitAsync(EngineName);
+            await client.CreateDatabaseAsync(Dbname, EngineName);
 
             var query = "x, x^2, x^3, x^4 from x in {1; 2; 3; 4; 5}";
-            var rsp = client.Execute(Dbname, EngineName, query, true);
+            var rsp = await client.ExecuteWaitAsync(Dbname, EngineName, query, true);
 
             var results = new List<ArrowRelation>
             {
@@ -39,12 +40,12 @@ namespace RelationalAI.Test
             Assert.Equal(rsp.Problems, problems);
         }
 
-        public override void Dispose()
+        public override async Task DisposeAsync()
         {
             var client = CreateClient();
 
-            try { client.DeleteDatabase(Dbname); } catch { }
-            try { client.DeleteEngineWait(EngineName); } catch { }
+            try { await client.DeleteDatabaseAsync(Dbname); } catch {}
+            try { await client.DeleteEngineWaitAsync(EngineName); } catch {}
         }
     }
 }